requirements.txt
faiss_cpu==1.8.0.post1
fastapi==0.114.1
numpy==2.1.1
psycopg2_binary==2.9.9
pydantic==2.9.1
sentence_transformers==3.1.0
transformers==4.44.2
uvicorn==0.30.6


docker_image ==  ankane/pgvector
postgress 